Refactor: useCountdown, calculateTime으로 리팩토링
